+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/implementation/converters/ChatMessageConverter.java
@@ -28,7 +28,8 @@ public final class ChatMessageConverter {
             .setCreatedOn(obj.getCreatedOn())
             .setDeletedOn(obj.getDeletedOn())
             .setEditedOn(obj.getEditedOn())
-            .setSenderDisplayName(obj.getSenderDisplayName());
+            .setSenderDisplayName(obj.getSenderDisplayName())
+            .setMetadata(obj.getMetadata());
 
         if (obj.getSenderCommunicationIdentifier() != null) {
             chatMessage.setSenderCommunicationIdentifier(